feat(mixer): offer a relaunch when the capture grant lands too late
macOS settles the audio-capture question once per process. A permission
granted while Somul is running is therefore invisible to every tap this
process will ever create, and the panel's "waiting for the grant" state
would sit there forever telling the user to be patient about something
that can no longer happen.

The backend now reports whether it has retried and got the same answer,
which is what separates a grant still on its way from one that needs a
new process. The notice offers waiting for the first and a relaunch for
the second, through `relaunch_app` — which releases the single-instance
lock before restarting, or the replacement would find the lock taken and
quit, leaving no Somul running at all.

+++ src-tauri/src/meter.rs
@@ -812,7 +812,7 @@ mod tests {
             Arc::clone(&emitter) as Arc<dyn PanelEventEmitter>,
         );
 
-        thread::sleep(TICK * (SESSION_POLL_EVERY_TICKS as u32 * 2 + 4));
+        thread::sleep(TICK * (SESSION_POLL_EVERY_TICKS * 2 + 4));
         loop_under_test.stop();
 
         assert_eq!(
